+++ test/mmdb2_decoder/pointer_test.exs
@@ -1,0 +1,17 @@
+defmodule MMDB2Decoder.PointerTest do
+  use ExUnit.Case, async: true
+
+  alias MMDB2Decoder.TestHelpers.Fixture
+
+  test "find_pointer!/3 equals find_pointer/3" do
+    {:ok, meta, tree, _} =
+      :fixture_ipv4_24
+      |> Fixture.contents()
+      |> MMDB2Decoder.parse_database()
+
+    result_bang = MMDB2Decoder.find_pointer!({1, 1, 1, 3}, meta, tree)
+    result_regular = MMDB2Decoder.find_pointer({1, 1, 1, 3}, meta, tree)
+
+    assert {:ok, result_bang} == result_regular
+  end
+end